Add files via upload
This text file contains my operation history for OpenRefine. It is in a JSON format. In OpenRefine the code can be applied to the Aarhus citizenship record from 1740 to 1862, which has been provided by Aarhus Stadsarkiv. For access to the original dataset, please refer to the GitHub repository belonging to Aarhus Stadsarkiv: https://github.com/aarhusstadsarkiv/datasets/tree/master/citizenship/1740-1862.